Listado de las altas y cambios varios

+++ src/main/java/clases/Administrador.java
@@ -17,7 +17,7 @@ public class Administrador extends Persona implements java.io.Serializable {
 		
 	}
 	
-	public Administrador(String nombre, String apellido, Date fechaNacimiento, double dni, String email, int rol, String usuario, String contraseña){
+	public Administrador(String nombre, String apellido, Date fechaNacimiento, long dni, String email, int rol, String usuario, String contraseña){
 		super(nombre, apellido, fechaNacimiento, dni, email, rol, usuario, contraseña);
 		this.publicaciones = new ArrayList<Publicacion>();
 		this.comentarios = new ArrayList<Comentario>();
