+++ src/main/java/interfacesDAO/ProfesorDAO.java
@@ -7,6 +7,6 @@ import clases.*;
 
 public interface ProfesorDAO extends GenericDAO<Profesor> {
 	List<Publicacion> obtenerPublicaciones(long id);
-	List<Cartelera> obtenerCarteleras(long id);
+	List<Cartelera> obtenerCarteleras(int id);
 	List<Comentario> obtenerComentarios(long id);
 }
